+++ web/src/components/PaperSheet/index.js
@@ -1,0 +1,1 @@
+export { default } from './PaperSheet'
