+++ web/src/components/PropertyForm/PropertyFormStyles.js
@@ -1,0 +1,7 @@
+export const getFormStyles = () => {
+  return (theme) => ({
+    root: {
+      margin: theme.spacing.unit
+    }
+  })
+}
