+++ web/src/constants/assets.js
@@ -1,0 +1,1 @@
+export const APP_ICON = '/assets/icon.png'
