+++ web/src/constants/labels.js
@@ -1,0 +1,4 @@
+export const APP_TITLE = 'Property Reporter'
+export const DASHBOARD_PAGE_TITLE = 'Dashboard'
+export const SUMMARY_PAGE_TITLE = 'Property Summary'
+export const LIST_PAGE_TITLE = 'Property List'
