+++ web/src/containers/App/App.js
@@ -4,6 +4,13 @@ import AppRoot from 'components/AppRoot'
 import DashboardPage from 'containers/DashboardPage'
 import SummaryPage from 'containers/SummaryPage'
 import ListPage from 'containers/ListPage'
+import FormPage from 'containers/FormPage'
+
+import {
+  DASHBOARD_PAGE_TITLE,
+  SUMMARY_PAGE_TITLE,
+  LIST_PAGE_TITLE
+} from 'constants/labels'
 
 const App = () => {
   return (
@@ -11,16 +18,17 @@ const App = () => {
       <AppRoot
         links={
           [
-            { href: "/", label: "Dashboard" },
-            { href: "/list", label: "Property List" },
-            { href: "/summary", label: "Property Summary" }
+            { href: "/", label: DASHBOARD_PAGE_TITLE },
+            { href: "/summary", label: SUMMARY_PAGE_TITLE },
+            { href: "/list", label: LIST_PAGE_TITLE }
           ]
         }
         navs={
           [
             { href: "/", component: DashboardPage },
+            { href: "/summary", component: SummaryPage },
             { href: "/list", component: ListPage },
-            { href: "/summary", component: SummaryPage }
+            { href: "/detail", component: FormPage }
           ]
         }
       />
